# docs/swagger.yaml
basePath: /x_ray
definitions:
...
  dto.CreateInterviewReq:
    properties:
      company_id:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/dto.Difficulty'
      experience:
        type: integer
      got_offer:
        type: boolean
      job_title:
        type: string
      user_id:
        type: string
    type: object
...
  dto.Difficulty:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Easy
    - Medium
    - Hard
    - VeryHard
  dto.InterviewRes:
    properties:
      company_id:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/dto.Difficulty'
      experience:
        type: integer
      got_offer:
        type: boolean
      id:
        type: string
      job_title:
        type: string
    type: object
...
  localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      time:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: this Backend is for x-ray, a company review platform. this backend/project
    was build with the intention to providing transparency to the job seekers, and
    also to give the company a chance to improve their work environment. this project
    is still in early stage, so expect some breaking changes in the future.
  title: X-ray API
  version: "1.0"
paths:
...
  /interview:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of interviews with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Interviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.InterviewRes'
                  type: array
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Interviews by Pagination
      tags:
      - Interview
    post:
      consumes:
      - application/json
      description: Creates a new interview experience for a company (requires authentication)
      parameters:
      - description: 'Interview details (difficulty: 1=Easy, 2=Medium, 3=Hard, 4=VeryHard)'
        in: body
        name: interview
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInterviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: Interview created successfully
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "400":
          description: Invalid request body or difficulty value
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create Interview
      tags:
      - Interview
  /interview/{interview_id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific interview experience
      parameters:
      - description: Interview ID (UUID)
        in: path
        name: interview_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Interview fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.InterviewRes'
              type: object
        "400":
          description: Missing interview_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: Interview not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Interview by ID
      tags:
      - Interview
  /interview/company/{company_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all interview experiences for a specific company with
        pagination
      parameters:
      - description: Company ID (UUID)
        in: path
        name: company_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Interviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.InterviewRes'
                  type: array
              type: object
        "400":
          description: Missing company_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Interviews by Company ID
      tags:
      - Interview
  /interview/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all interview experiences submitted by a specific user
        with pagination
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Interviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.InterviewRes'
                  type: array
              type: object
        "400":
          description: Missing user_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Interviews by User ID
      tags:
      - Interview
...
swagger: "2.0"